+++ lib/re2.dart
@@ -23,3 +23,4 @@ library;
 
 export 'src/re2_base.dart' show Re2;
 export 'src/re2_match.dart' show Re2Match;
+export 'src/re2_set.dart' show Re2Set;